# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Labyrinths — server-authoritative high-fantasy dungeon economy game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /labyrinths:
    get:
      operationId: listLabyrinths
      tags: [labyrinths]
      summary: Browse the published overworld of labyrinths
      parameters:
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [trending, top_rated, newest, most_run, highest_reward]
        - name: entry
          in: query
          required: false
          schema:
            type: string
            enum: [all, free, paid]
        - name: biome
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Published labyrinths
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Labyrinth"
  /labyrinths/claim:
    post:
      operationId: claimLabyrinth
      tags: [labyrinths]
      summary: Claim a fresh labyrinth for the current player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabyrinthClaimInput"
      responses:
        "201":
          description: Claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Labyrinth"
        "409":
          description: Player already owns a labyrinth
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /labyrinths/{id}:
    get:
      operationId: getLabyrinth
      tags: [labyrinths]
      summary: Get a labyrinth by id
      parameters:
        - $ref: "#/components/parameters/LabyrinthId"
      responses:
        "200":
          description: Labyrinth
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Labyrinth"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateLabyrinth
      tags: [labyrinths]
      summary: Update labyrinth name, description, or theme
      parameters:
        - $ref: "#/components/parameters/LabyrinthId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabyrinthUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Labyrinth"
  /labyrinths/{id}/preview:
    get:
      operationId: getLabyrinthPreview
      tags: [labyrinths]
      summary: Get a rich preview (chambers, loot table, difficulty) before running
      parameters:
        - $ref: "#/components/parameters/LabyrinthId"
      responses:
        "200":
          description: Preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabyrinthPreview"
  /labyrinths/{id}/publish:
    post:
      operationId: publishLabyrinth
      tags: [labyrinths]
      summary: Publish a labyrinth to the overworld
      parameters:
        - $ref: "#/components/parameters/LabyrinthId"
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Labyrinth"
  /labyrinths/{id}/unpublish:
    post:
      operationId: unpublishLabyrinth
      tags: [labyrinths]
      summary: Remove a labyrinth from the overworld
      parameters:
        - $ref: "#/components/parameters/LabyrinthId"
      responses:
        "200":
          description: Unpublished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Labyrinth"
  /labyrinths/{id}/ratings:
    get:
      operationId: listLabyrinthRatings
      tags: [labyrinths]
      summary: List ratings for a labyrinth
      parameters:
        - $ref: "#/components/parameters/LabyrinthId"
      responses:
        "200":
          description: Ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rating"
...
components:
  parameters:
    LabyrinthId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LabyrinthClaimInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        biome:
          type: string
    LabyrinthUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        biome:
          type: string
    Labyrinth:
      type: object
      required:
        - id
        - ownerUserId
        - ownerName
        - ownerAvatarUrl
        - name
        - description
        - biome
        - level
        - depth
        - chamberCount
        - rareNodeCount
        - published
        - tollGateUnlocked
        - entryFee
        - ratingAverage
        - ratingCount
        - runsToday
        - runsAllTime
        - rewardValueToday
        - dailyRunCapacity
        - dailyRewardCapacity
        - appealScore
        - bossActive
        - isOwner
        - createdAt
      properties:
        id:
          type: integer
        ownerUserId:
          type: integer
        ownerName:
          type: string
        ownerAvatarUrl:
          type: string
        name:
          type: string
        description:
          type: string
        biome:
          type: string
        level:
          type: integer
        depth:
          type: integer
        chamberCount:
          type: integer
        rareNodeCount:
          type: integer
        published:
          type: boolean
        tollGateUnlocked:
          type: boolean
        entryFee:
          type: integer
        ratingAverage:
          type: number
        ratingCount:
          type: integer
        runsToday:
          type: integer
        runsAllTime:
          type: integer
        rewardValueToday:
          type: integer
        dailyRunCapacity:
          type: integer
        dailyRewardCapacity:
          type: integer
        appealScore:
          type: integer
        bossActive:
          type: boolean
        isOwner:
          type: boolean
        accentColor:
          type: string
        createdAt:
          type: string
...
    LootTableEntry:
      type: object
      required: [kind, label, rarity, chance, minValue, maxValue]
      properties:
        kind:
          type: string
          enum: [currency, material, item]
        label:
          type: string
        rarity:
          type: string
          enum: [common, uncommon, rare, epic, legendary]
        chance:
          type: number
        minValue:
          type: integer
        maxValue:
          type: integer
        icon:
          type: string
    LabyrinthPreview:
      type: object
      required:
        - labyrinth
        - chambers
        - lootTable
        - estimatedClearSeconds
        - difficulty
      properties:
        labyrinth:
          $ref: "#/components/schemas/Labyrinth"
        chambers:
          type: array
          items:
            $ref: "#/components/schemas/ChamberSummary"
        lootTable:
          type: array
          items:
            $ref: "#/components/schemas/LootTableEntry"
        estimatedClearSeconds:
          type: integer
        difficulty:
          type: string
          enum: [novice, adept, veteran, master, mythic]
    ChamberSummary:
      type: object
      required: [id, name, biome, difficulty, enemyCount, hasBoss, lootTier]
      properties:
        id:
          type: integer
        name:
          type: string
        biome:
          type: string
        difficulty:
          type: integer
        enemyCount:
          type: integer
        hasBoss:
          type: boolean
        lootTier:
          type: integer
    Rating:
      type: object
      required: [id, labyrinthId, raterName, raterAvatarUrl, stars, createdAt]
      properties:
        id:
          type: integer
        labyrinthId:
          type: integer
        raterName:
          type: string
        raterAvatarUrl:
          type: string
        stars:
          type: integer
        comment:
          type: ["string", "null"]
        difficultyVote:
          type: ["string", "null"]
        createdAt:
          type: string
...